+++ server/src/chattingRoom/routes.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const chattingRoomController = require("./chattingRoomController");
+
+router.post("/", chattingRoomController);
+
+module.exports = router;
